Gender Name Classification/requirements.txt
streamlit==0.67.1
pandas==1.1.2
numpy==1.19.2
scikit-learn==0.23.2
joblib==0.16.0
Pillow==6.2.0
matplotlib==3.1.1